import tarball aegis-2008.03.09-04.024

+++ libaegis/input/gunzip.cc
@@ -1,10 +1,10 @@
 //
 //	aegis - project change supervisor
-//	Copyright (C) 1999, 2001-2007 Peter Miller
+//	Copyright (C) 1999, 2001-2008 Peter Miller
 //
 //	This program is free software; you can redistribute it and/or modify
 //	it under the terms of the GNU General Public License as published by
-//	the Free Software Foundation; either version 2 of the License, or
+//	the Free Software Foundation; either version 3 of the License, or
 //	(at your option) any later version.
 //
 //	This program is distributed in the hope that it will be useful,
@@ -13,15 +13,14 @@
 //	GNU General Public License for more details.
 //
 //	You should have received a copy of the GNU General Public License
-//	along with this program; if not, write to the Free Software
-//	Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111, USA.
-//
-// MANIFEST: functions to manipulate gunzips
+//	along with this program.  If not, see
+//	<http://www.gnu.org/licenses/>.
 //
 
 #include <common/ac/zlib.h>
 #include <common/ac/string.h>
 
+#include <common/mem.h>
 #include <libaegis/input/gunzip.h>
 #include <libaegis/sub.h>
 
@@ -189,7 +188,7 @@ input_gunzip::read_inner(void *data, size_t len)
 	// Watch out for 64-bit machines.  This is what
 	// those aparrently redundant 0xFFFFFFFF are for.
 	//
-	if ((getLong() & 0xFFFFFFFF) != (crc & 0xFFFFFFFF))
+	if (((unsigned long)getLong() & 0xFFFFFFFF) != (crc & 0xFFFFFFFF))
 	    fatal_error("gunzip: checksum mismatch");
 
 	//
